chore: minor movement of tests among files
Moves tests to more relevant test file.
