+++ tests/components/solax/test_init.py
@@ -2,24 +2,17 @@
 
 from unittest.mock import patch
 
-from freezegun.api import FrozenDateTimeFactory
 import pytest
-from solax.inverter import InverterError, InverterResponse
+from solax.inverter import InverterResponse
 from solax.inverters import X1MiniV34
 
-from homeassistant.components.solax import SCAN_INTERVAL
 from homeassistant.components.solax.const import DOMAIN
 from homeassistant.config_entries import ConfigEntryState
-from homeassistant.const import (
-    CONF_IP_ADDRESS,
-    CONF_MODEL,
-    CONF_PASSWORD,
-    CONF_PORT,
-    STATE_UNAVAILABLE,
-)
+from homeassistant.const import CONF_IP_ADDRESS, CONF_MODEL, CONF_PASSWORD, CONF_PORT
 from homeassistant.core import HomeAssistant
+from homeassistant.helpers import device_registry as dr
 
-from tests.common import MockConfigEntry, async_fire_time_changed
+from tests.common import MockConfigEntry
 
 
 def __mock_get_data() -> InverterResponse:
@@ -84,12 +77,12 @@ async def test_setup_entry_success(
     assert mock_discover.call_args.kwargs.get("inverters") == expected_inverters
 
 
-async def test_coordinator_update_failure(
+async def test_device_info_model(
     hass: HomeAssistant,
-    freezer: FrozenDateTimeFactory,
+    device_registry: dr.DeviceRegistry,
     mock_config_entry: MockConfigEntry,
 ) -> None:
-    """Test an InverterError during a refresh marks the entities unavailable."""
+    """Test the device registry entry has the discovered inverter model set."""
     mock_config_entry.add_to_hass(hass)
 
     inverter = next(
@@ -109,23 +102,9 @@ async def test_coordinator_update_failure(
         assert await hass.config_entries.async_setup(mock_config_entry.entry_id)
         await hass.async_block_till_done()
 
-    assert mock_config_entry.state is ConfigEntryState.LOADED
-
-    # making sure the base state is actually healthy before the test run.
-    assert (
-        hass.states.get("sensor.solax_abcdefghij_network_voltage").state
-        != STATE_UNAVAILABLE
-    )
-
-    with patch("solax.RealTimeAPI.get_data", side_effect=InverterError):
-        freezer.tick(SCAN_INTERVAL)
-        async_fire_time_changed(hass)
-        await hass.async_block_till_done()
-
-    assert (
-        hass.states.get("sensor.solax_abcdefghij_network_voltage").state
-        == STATE_UNAVAILABLE
-    )
+    device = device_registry.async_get_device(identifiers={(DOMAIN, "ABCDEFGHIJ")})
+    assert device is not None
+    assert device.model == "x1_mini_v34"
 
 
 async def test_setup_entry_not_ready(
